+++ src/sql/plot.py
@@ -24,7 +24,6 @@ except ModuleNotFoundError:
 import sql.connection
 from sql.telemetry import telemetry
 import warnings
-from sql import util
 
 
 def _summary_stats(conn, table, column, with_=None):
@@ -140,9 +139,6 @@ def _boxplot_stats(conn, table, column, whis=1.5, autorange=False, with_=None):
     """Compute statistics required to create a boxplot"""
     if not conn:
         conn = sql.connection.ConnectionManager.current
-
-    # calculating stats might fail on other DBs (percentile_disc)
-    util.support_only_sql_alchemy_connection("boxplot")
 
     def _compute_conf_interval(N, med, iqr):
         notch_min = med - 1.57 * iqr / np.sqrt(N)